estoy con puts

+++ src/main/java/com/example/tfg/repositorios/RutinaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.tfg.modelos.Rutina;
 
-public interface RutinaRepository extends JpaRepository<Rutina, Integer>{
+public interface RutinaRepository extends JpaRepository<Rutina, Long>{
 
 }
